IS: Refactor domain classes into one domain folder
