+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/AktivitetskravService.kt
@@ -3,10 +3,18 @@ package no.nav.syfo.aktivitetskrav
 import no.nav.syfo.aktivitetskrav.api.DocumentComponentDTO
 import no.nav.syfo.aktivitetskrav.api.HistorikkDTO
 import no.nav.syfo.aktivitetskrav.api.createHistorikkDTOs
-import no.nav.syfo.aktivitetskrav.domain.*
 import no.nav.syfo.infrastructure.kafka.AktivitetskravVurderingProducer
 import no.nav.syfo.application.exception.ConflictException
+import no.nav.syfo.domain.Aktivitetskrav
+import no.nav.syfo.domain.AktivitetskravStatus
+import no.nav.syfo.domain.AktivitetskravVarsel
+import no.nav.syfo.domain.AktivitetskravVurdering
 import no.nav.syfo.domain.PersonIdent
+import no.nav.syfo.domain.isInFinalState
+import no.nav.syfo.domain.oppfyllAutomatisk
+import no.nav.syfo.domain.toVarselType
+import no.nav.syfo.domain.updateStoppunkt
+import no.nav.syfo.domain.validate
 import no.nav.syfo.infrastructure.database.repository.updateAktivitetskrav
 import no.nav.syfo.oppfolgingstilfelle.domain.Oppfolgingstilfelle
 import java.sql.Connection
