+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/AktivitetskravVarselService.kt
@@ -2,11 +2,16 @@ package no.nav.syfo.aktivitetskrav
 
 import no.nav.syfo.aktivitetskrav.api.ForhandsvarselDTO
 import no.nav.syfo.infrastructure.database.repository.AktivitetskravVarselRepository
-import no.nav.syfo.aktivitetskrav.domain.*
 import no.nav.syfo.application.exception.ConflictException
+import no.nav.syfo.domain.Aktivitetskrav
+import no.nav.syfo.domain.AktivitetskravStatus
+import no.nav.syfo.domain.AktivitetskravVarsel
+import no.nav.syfo.domain.AktivitetskravVurdering
 import no.nav.syfo.infrastructure.kafka.domain.ExpiredVarsel
 import no.nav.syfo.infrastructure.kafka.domain.KafkaAktivitetskravVarsel
 import no.nav.syfo.domain.PersonIdent
+import no.nav.syfo.domain.VarselType
+import no.nav.syfo.domain.isInFinalState
 import no.nav.syfo.infrastructure.kafka.AktivitetskravVarselProducer
 import no.nav.syfo.infrastructure.kafka.AktivitetskravVurderingProducer
 import no.nav.syfo.infrastructure.kafka.ExpiredVarselProducer
