+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/IAktivitetskravVarselRepository.kt
@@ -2,9 +2,9 @@ package no.nav.syfo.aktivitetskrav
 
 import no.nav.syfo.infrastructure.database.repository.PAktivitetskravVarsel
 import no.nav.syfo.infrastructure.database.repository.VarselReferences
-import no.nav.syfo.aktivitetskrav.domain.Aktivitetskrav
-import no.nav.syfo.aktivitetskrav.domain.AktivitetskravVarsel
-import no.nav.syfo.aktivitetskrav.domain.AktivitetskravVurdering
+import no.nav.syfo.domain.Aktivitetskrav
+import no.nav.syfo.domain.AktivitetskravVarsel
+import no.nav.syfo.domain.AktivitetskravVurdering
 import no.nav.syfo.infrastructure.kafka.domain.ExpiredVarsel
 import no.nav.syfo.infrastructure.kafka.domain.KafkaAktivitetskravVarsel
 import no.nav.syfo.domain.PersonIdent
