+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/VarselPdfService.kt
@@ -1,7 +1,7 @@
 package no.nav.syfo.aktivitetskrav
 
-import no.nav.syfo.aktivitetskrav.domain.AktivitetskravVarsel
-import no.nav.syfo.aktivitetskrav.domain.VarselType
+import no.nav.syfo.domain.AktivitetskravVarsel
+import no.nav.syfo.domain.VarselType
 import no.nav.syfo.client.pdfgen.ForhandsvarselPdfDTO
 import no.nav.syfo.client.pdfgen.PdfGenClient
 import no.nav.syfo.client.pdfgen.VurderingPdfDTO
@@ -15,7 +15,7 @@ class VarselPdfService(
     suspend fun createVarselPdf(
         personIdent: PersonIdent,
         varsel: AktivitetskravVarsel,
-        callId: String
+        callId: String,
     ): ByteArray = when (varsel.type) {
         VarselType.FORHANDSVARSEL_STANS_AV_SYKEPENGER -> {
             val personNavn = pdlClient.navn(personIdent)
