+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/cronjob/AktivitetskravNyCronjob.kt
@@ -2,7 +2,8 @@ package no.nav.syfo.aktivitetskrav.cronjob
 
 import net.logstash.logback.argument.StructuredArguments
 import no.nav.syfo.aktivitetskrav.AktivitetskravService
-import no.nav.syfo.aktivitetskrav.domain.*
+import no.nav.syfo.domain.AktivitetskravStatus
+import no.nav.syfo.domain.isAutomatiskOppfylt
 import no.nav.syfo.infrastructure.cronjob.Cronjob
 import no.nav.syfo.infrastructure.cronjob.CronjobResult
 import no.nav.syfo.infrastructure.database.DatabaseInterface
