+++ src/main/kotlin/no/nav/syfo/infrastructure/database/repository/AktivitetskravRepository.kt
@@ -3,9 +3,9 @@ package no.nav.syfo.infrastructure.database.repository
 import com.fasterxml.jackson.core.type.TypeReference
 import no.nav.syfo.aktivitetskrav.IAktivitetskravRepository
 import no.nav.syfo.aktivitetskrav.api.DocumentComponentDTO
-import no.nav.syfo.aktivitetskrav.domain.Aktivitetskrav
-import no.nav.syfo.aktivitetskrav.domain.AktivitetskravStatus
-import no.nav.syfo.aktivitetskrav.domain.AktivitetskravVurdering
+import no.nav.syfo.domain.Aktivitetskrav
+import no.nav.syfo.domain.AktivitetskravStatus
+import no.nav.syfo.domain.AktivitetskravVurdering
 import no.nav.syfo.infrastructure.database.DatabaseInterface
 import no.nav.syfo.infrastructure.database.NoElementInsertedException
 import no.nav.syfo.infrastructure.database.toList
