+++ src/main/kotlin/no/nav/syfo/infrastructure/kafka/AktivitetskravVurderingProducer.kt
@@ -1,7 +1,7 @@
 package no.nav.syfo.infrastructure.kafka
 
-import no.nav.syfo.aktivitetskrav.domain.Aktivitetskrav
-import no.nav.syfo.aktivitetskrav.domain.AktivitetskravStatus
+import no.nav.syfo.domain.Aktivitetskrav
+import no.nav.syfo.domain.AktivitetskravStatus
 import no.nav.syfo.infrastructure.kafka.domain.AktivitetskravVurderingRecord
 import no.nav.syfo.util.configuredJacksonMapper
 import org.apache.kafka.clients.producer.KafkaProducer
