+++ src/main/kotlin/no/nav/syfo/infrastructure/kafka/domain/ExpiredVarsel.kt
@@ -1,6 +1,6 @@
 package no.nav.syfo.infrastructure.kafka.domain
 
-import no.nav.syfo.aktivitetskrav.domain.VarselType
+import no.nav.syfo.domain.VarselType
 import no.nav.syfo.domain.PersonIdent
 import java.time.LocalDate
 import java.time.LocalDateTime
