+++ src/test/kotlin/no/nav/syfo/aktivitetskrav/cronjob/PubliserAktivitetskravVarselCronjobSpek.kt
@@ -8,7 +8,7 @@ import no.nav.syfo.aktivitetskrav.VarselPdfService
 import no.nav.syfo.aktivitetskrav.api.DocumentComponentDTO
 import no.nav.syfo.infrastructure.database.repository.AktivitetskravRepository
 import no.nav.syfo.infrastructure.database.repository.AktivitetskravVarselRepository
-import no.nav.syfo.aktivitetskrav.domain.*
+import no.nav.syfo.domain.*
 import no.nav.syfo.infrastructure.kafka.AktivitetskravVarselProducer
 import no.nav.syfo.infrastructure.kafka.domain.KafkaAktivitetskravVarsel
 import no.nav.syfo.testhelper.ExternalMockEnvironment
